BeforeSuggestioEvent and SuggestionEvent moved to Service namespace
