+++ src/contracts/Event/Service/BeforeSuggestEvent.php
@@ -6,7 +6,7 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\Contracts\Search\Event;
+namespace Ibexa\Contracts\Search\Event\Service;
 
 use Ibexa\Contracts\Core\Repository\Event\BeforeEvent;
 use Ibexa\Contracts\Search\Model\Suggestion\SuggestionCollection;
